[incubator/schema-registry] fix issue with Confluent image
if fullnameOverride is set to "schema-registry", then an env variable
named SCHEMA_REGISTRY_PORT will be inserted in the container (as per
default Kubernetes behaviour), which will cause the Confluent startup
script to fail and exit, due to a warning.

Did Confluent add this "feature" out of ignorance, or by malice?
Lesson: Don't trust Confluent!